+++ front-end/src/App.jsx
@@ -2,9 +2,9 @@ import React from 'react';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { BrowserRouter, Route, Routes, useLocation } from 'react-router';
 import 'react-datepicker/dist/react-datepicker.css';
-import './index.css';
-import './customDatePicker.css';
-import Layout from './Layout';
+import './styles/index.css';
+import './styles/customDatePicker.css';
+import Layout from './components/Layout';
 import FeedbackRequest from './pages/feedback/FeedbackRequest';
 import Splash from './pages/auth/Splash';
 import SignIn from './pages/auth/SignIn';
@@ -20,13 +20,13 @@ import TeamSpaceManage from './pages/teamspace/TeamSpaceManage';
 import TeamSpaceEdit from './pages/teamspace/TeamSpaceEdit';
 import FeedbackComplete from './pages/feedback/FeedbackComplete';
 import FeedbackSelf from './pages/feedback/FeedbackSelf';
-import CombinedProvider from './CombinedProvider';
+import CombinedProvider from './store/CombinedProvider';
 import FeedbackSendLayout from './pages/feedback/FeedbackSendLayout';
 import FeedbackSendStep from './pages/feedback/FeedbackSendStep';
 import FeedbackSend from './pages/feedback/FeedbackSend';
 import FeedbackFavorite from './pages/feedback/FeedbackFavorite';
 import MyPageHome from './pages/mypage/MyPageHome';
-import ProtectedRoute from './ProtectedRoute';
+import ProtectedRoute from './components/ProtectedRoute';
 import SplashForOAuth from './pages/auth/SplashForOAuth';
 import FeedbackSendFreq from './pages/feedback/FeedbackSendFreq';
 import ProfileEdit from './pages/mypage/ProfileEdit';
