+++ front-end/src/components/Layout.jsx
@@ -1,6 +1,6 @@
 import { Outlet, useLocation } from 'react-router-dom';
-import ModalBase from './components/modals/ModalBase';
-import ToastBase from './components/toasts/ToastBase';
+import ModalBase from './modals/ModalBase';
+import ToastBase from './toasts/ToastBase';
 
 const noPaddingPages = ['/main', '/calendar']; // 패딩을 제거할 페이지 리스트
