+++ front-end/src/styles/slider.css
