+++ front-end/src/utility/useHandlePop.jsx
@@ -1,5 +1,4 @@
 import { useEffect } from 'react';
-import { useNavigate } from 'react-router-dom';
 
 const useHandlePop = (handlePopState) => {
   useEffect(() => {
